requirements.txt
dj-database-url==0.5.0
Django==2.1.1
django-admin-rangefilter==0.3.7
django-heroku==0.3.1
djangorestframework==3.8.2
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.5
whitenoise==4.1
